+++ core/src/main/java/arez/SpyImpl.java
@@ -232,15 +232,6 @@ final class SpyImpl
    * {@inheritDoc}
    */
   @Override
-  public boolean isRunning( @Nonnull final Observer observer )
-  {
-    return isTransactionActive() && null != getTrackerTransaction( observer );
-  }
-
-  /**
-   * {@inheritDoc}
-   */
-  @Override
   public boolean isComputedValue( @Nonnull final ObservableValue<?> observableValue )
   {
     return observableValue.hasOwner();